+++ comics/comics/pinkparts.py
@@ -7,17 +7,11 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://pinkpartscomic.com/'
     start_date = '2010-02-01'
+    end_date = '2014-02-23'
+    active = False
     rights = 'Katherine Skipper'
 
 
 class Crawler(CrawlerBase):
-    history_capable_date = '2010-02-01'
-    time_zone = 'US/Pacific'
-
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://pinkpartscomic.com/inc/feed.php')
-        for entry in feed.for_date(pub_date):
-            page = self.parse_page(entry.link)
-            url = page.src('img[src*="/img/comic/"]')
-            title = entry.title.replace('New comic: ', '')
-            return CrawlerImage(url, title)
+        pass  # Comic no longer published